stupid github desktop

+++ Game/src/Engine/FrameTimer.hpp
@@ -16,11 +16,11 @@ class FrameTimer
 {
 public:
     FrameTimer();
-    FrameTimer(std::vector<Texture2D *> frames);
+    FrameTimer(std::vector<Texture2D*> frames);
     FrameTimer(std::vector<std::string> frames);
     FrameTimer(std::vector<std::filesystem::path> frames);
-    FrameTimer(std::vector<SDL_Texture *> frames);
-    FrameTimer(std::vector<Texture2D_Vulkan *> frames);
+    FrameTimer(std::vector<SDL_Texture*> frames);
+    FrameTimer(std::vector<Texture2D_Vulkan*> frames);
     ~FrameTimer();
 
     bool   Repeat;
@@ -35,8 +35,8 @@ public:
     Vector2 AbsoluteSize;
 
     void Draw(double delta);
-    void Draw(double delta, Rect *clip);
-    void SetFPS(float fps);
+    void Draw(double delta, Rect* clip);
+    void SetFPS(double fps);
     void ResetIndex();
     void LastIndex();
     void SetIndexAt(int idx);
